+++ Sumo/src/main/java/com/codestates/comment/service/CommentService.java
@@ -1,10 +1,11 @@
 package com.codestates.comment.service;
 
-
-
+import com.codestates.auth.LoginUtils;
 import com.codestates.board.entity.Board;
 import com.codestates.board.repository.BoardRepository;
 import com.codestates.comment.entity.Comment;
+import com.codestates.comment.entity.CommentLike;
+import com.codestates.comment.repository.CommentLikesRepository;
 import com.codestates.comment.repository.CommentRepository;
 import com.codestates.exception.BusinessLogicException;
 import com.codestates.exception.ExceptionCode;
@@ -13,11 +14,9 @@ import com.codestates.member.repository.MemberRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
-
 
 
 @Service
@@ -33,28 +32,38 @@ public class CommentService {
     @Autowired
     private BoardRepository boardRepository;
 
+    @Autowired
+    private CommentLikesRepository commentLikesRepository;
+
+
+
     //댓글 생성
-    //Todo : Security 적용시 주석해제
     @Transactional
-    public Comment createComment(Comment comment){
-//        Member currentMember = getCurrentMember();
-//        comment.setMember(currentMember);
+    public Comment createComment(Comment comment, Long boardId){
+        Member currentMember = getCurrentMember();
+        comment.setMember(currentMember);
+
+        Board board = boardRepository.findById(boardId)
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.BOARD_NOT_FOUND));
+
+        comment.setBoard(board);
+        board.addComment(comment);
+        boardRepository.save(board); //있어도, 없어도됨 @Transactional 있음.
 
         return commentRepository.save(comment);
     }
 
     //댓글 수정
-    //Todo : Security 적용시 주석해제
     @Transactional
     public Comment updateComment(Comment comment){
-//        Member currentMember = getCurrentMember();
 
+        Member currentMember = getCurrentMember();
         Comment findComment = findVerifiedComment(comment.getCommentId());
 
-//        if(!findComment.getMember().getMemberId().equals(currentMember.getMemberId())) {
-//            throw new BusinessLogicException(ExceptionCode.COMMENT_ACCESS_DENIED);
-//
-//        }
+        if(!findComment.getMember().getMemberId().equals(currentMember.getMemberId())) {
+            throw new BusinessLogicException(ExceptionCode.COMMENT_ACCESS_DENIED);
+
+        }
         Optional.ofNullable(comment.getCommentContent()).ifPresent(commentContent -> findComment.setCommentContent(commentContent));
 
         findComment.setModifiedAt(LocalDateTime.now());
@@ -62,19 +71,16 @@ public class CommentService {
     }
 
     //게시글 삭제
-    //Todo : Security 적용시 주석 해제.
     public void deleteComment(long commentId){
-//        Member currentMember = getCurrentMember();
 
+        Member currentMember = getCurrentMember();
         Comment findComment = findVerifiedComment(commentId);
 
-//        if(!findComment.getMember().getMemberId().equals(currentMember.getMemberId())) {
-//            throw new BusinessLogicException(ExceptionCode.COMMENT_ACCESS_DENIED);
-//
-//        }
+        if(!findComment.getMember().getMemberId().equals(currentMember.getMemberId())) {
+            throw new BusinessLogicException(ExceptionCode.COMMENT_ACCESS_DENIED);
+        }
 
         commentRepository.deleteById(commentId);
-
     }
 
 
@@ -84,6 +90,7 @@ public class CommentService {
 
     //댓글 검증
     private Comment findVerifiedComment(long commentId){
+
         Optional<Comment> optionalComment = commentRepository.findById(commentId);
         Comment findComment =
                 optionalComment.orElseThrow(() ->
@@ -91,14 +98,15 @@ public class CommentService {
         return findComment;
     }
 
-    // TODO: 현재 로그인한 회원 정보 가지고오기. // // TODO: SECURITY 적용시 주석해제
-//    public Member getCurrentMember() {
-//            private Member getCurrentMember() {
-//        String email = SecurityContextHolder.getContext().getAuthentication().getName();
-//        return memberRepository.findByEmail(email)
-//                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
-//
-//    }
+
+    public Member getCurrentMember() {
+
+        String email = LoginUtils.checkLogin();
+
+        return memberRepository.findByEmail(email)
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
+    }
+
 
     // boardId를 통해서 comment리스트 목록 뽑아오기
     @Transactional(readOnly = true)
@@ -109,4 +117,37 @@ public class CommentService {
 
         return commentRepository.findByBoard(board);
     }
+
+
+    public void toggleLike(Long memberId, Long commentId){
+        Comment comment = findVerifiedComment(commentId);
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
+
+        Optional<CommentLike> commentLike = commentLikesRepository.findByCommentAndMember(comment, member);
+
+        if(commentLike.isPresent()) {
+            if (commentLike.get().getCommentLikeStatus() == CommentLike.CommentLikeStatus.LIKE){
+                commentLike.get().setCommentLikeStatus(CommentLike.CommentLikeStatus.DISLIKE);
+            } else {
+                commentLike.get().setCommentLikeStatus(CommentLike.CommentLikeStatus.LIKE);
+            }
+            commentLikesRepository.save(commentLike.get());
+        } else{
+            CommentLike newCommentLike = new CommentLike(comment, member);
+            newCommentLike.setCommentLikeStatus(CommentLike.CommentLikeStatus.LIKE);
+            commentLikesRepository.save(newCommentLike);
+        }
+
+        comment.setCommentLike(commentLikesRepository.findByComment(comment));
+        commentRepository.save(comment);
+    }
+
+
+    public int getCommentLikeCount(long commentId){
+        Comment comment = findVerifiedComment(commentId);
+        int commentLike = comment.getCommentLikeCount();
+        return commentLike;
+    }
+
 }
